Add SCSS styling configuration and SASS library
This commit adds _reset.scss, _values.scss, and other SCSS files to set up styling configurations for the application. Furthermore, SASS dependency is added to package.json as it is required for the SCSS preprocessing. These additions are crucial for creating a uniform look and feel across the web application while also providing easy maintenance of styles.