Removed unnecessarily complex CSS, updated routing, added react-helmet
The major changes in this commit are the removal of extensive CSS styles from client/src/App.css and client/src/index.css, replacing them with a simple SCSS file. Additionally, routing was updated to use react-router-dom's BrowserRouter, Route, and Routes components. Lastly, react-helmet was added to the server to better support SEO.