Added language support to the application
The changes include the creation of new translation files in the locales folder and the use of 'react-i18next', which is a powerful internationalization framework for React applications. Some components like Register, Login, and Page-Footer are updated with the localization functionality. Also, updates are made on the CSS files and adjustments in the variables names in 'settings.js'.