Add chat style functionality to Whispers page
A new state named 'activeStyle' has been added to the Whispers component to manage the currently selected chat style. A new component 'ChatStyle' was also introduced, allowing users to select from different chat styles. This selected style is then passed to the 'MessageList' component to update the appearance of the chat based on the user's selection.